requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pydantic==1.10.7